requirements.txt
pandas==2.2.3
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pyarrow==23.0.1